De-couple VideoTable._ID from GiantBomb ids, and make it auto-incrementing.
This is done because the games <=> videos relation is supposed to be
many-to-many, but I don't want to go to the trouble of creating a separate join
table for a case so rare as this. In the event that multiple games in the
library share a video, now I simply store multiple copies of the video in the
database.